+++ osp/osp.c
@@ -458,17 +458,36 @@ osp_delete_scan (osp_connection_t *connection, const char *scan_id)
  * @param[out]  graph       Graphic base64 encoded.
  * @param[out]  error       Pointer to error, if any.
  *
- * @return 0 if success, 1 if error.
+ * @return 0 if success, -1 if error.
  */
 int
 osp_get_performance_ext (osp_connection_t *connection,
-                         osp_get_performance_opts_t opts,
-                         char **graph, char **error)
+                         osp_get_performance_opts_t opts, char **graph,
+                         char **error)
 {
   entity_t entity;
   int rc;
+  time_t now;
 
-  assert (connection);
+  if (!connection)
+    {
+      if (error)
+        *error = g_strdup ("Couldn't send get_performance command "
+                           "to scanner. Not valid connection");
+      return -1;
+    }
+
+  time (&now);
+
+  if (!opts.titles || !strcmp (opts.titles, "") || opts.start < 0
+      || opts.start > now || opts.end < 0 || opts.end > now)
+    {
+      if (error)
+        *error = g_strdup ("Couldn't send get_performance command "
+                           "to scanner. Bad or missing parameters.");
+      return -1;
+    }
+
   rc = osp_send_command (connection, &entity,
                          "<get_performance start='%d' "
                          "end='%d' titles='%s'/>",
@@ -478,11 +497,11 @@ osp_get_performance_ext (osp_connection_t *connection,
     {
       if (error)
         *error = g_strdup ("Couldn't send get_performance command to scanner");
-      return 1;
+      return -1;
     }
 
   if (graph && entity_text (entity) && strcmp (entity_text (entity), "\0"))
-      *graph = g_strdup (entity_text (entity));
+    *graph = g_strdup (entity_text (entity));
   else
     {
       const char *text = entity_attribute (entity, "status_text");
@@ -491,7 +510,7 @@ osp_get_performance_ext (osp_connection_t *connection,
       if (error)
         *error = g_strdup (text);
       free_entity (entity);
-      return 1;
+      return -1;
     }
 
   free_entity (entity);
@@ -515,7 +534,14 @@ osp_get_scan_status_ext (osp_connection_t *connection,
   int rc;
   osp_scan_status_t status = OSP_SCAN_STATUS_ERROR;
 
-  assert (connection);
+  if (!connection)
+    {
+      if (error)
+        *error = g_strdup ("Couldn't send get_scans command "
+                           "to scanner. Not valid connection");
+      return status;
+    }
+
   assert (opts.scan_id);
   rc = osp_send_command (connection, &entity,
                          "<get_scans scan_id='%s'"
@@ -576,7 +602,13 @@ osp_get_scan_pop (osp_connection_t *connection, const char *scan_id,
   int progress;
   int rc;
 
-  assert (connection);
+  if (!connection)
+    {
+      if (error)
+        *error = g_strdup ("Couldn't send get_scan command "
+                           "to scanner. Not valid connection");
+      return -1;
+    }
   assert (scan_id);
   rc = osp_send_command (connection, &entity,
                          "<get_scans scan_id='%s'"
@@ -649,7 +681,13 @@ osp_stop_scan (osp_connection_t *connection, const char *scan_id, char **error)
   entity_t entity;
   int rc;
 
-  assert (connection);
+  if (!connection)
+    {
+      if (error)
+        *error = g_strdup ("Couldn't send stop_scan command "
+                           "to scanner. Not valid connection");
+      return -1;
+    }
   assert (scan_id);
   rc = osp_send_command (connection, &entity, "<stop_scan scan_id='%s'/>",
                          scan_id);
@@ -726,7 +764,14 @@ osp_start_scan (osp_connection_t *connection, const char *target,
   int status;
   int rc;
 
-  assert (connection);
+  if (!connection)
+    {
+      if (error)
+        *error = g_strdup ("Couldn't send start_scan command "
+                           "to scanner. Not valid connection");
+      return -1;
+    }
+
   assert (target);
   /* Construct options string. */
   if (options)
@@ -903,6 +948,14 @@ osp_start_scan_ext (osp_connection_t *connection,
   gchar *cmd;
   char filename[] = "/tmp/osp-cmd-XXXXXX";
   int fd;
+
+  if (!connection)
+    {
+      if (error)
+        *error = g_strdup ("Couldn't send start_scan command "
+                           "to scanner. Not valid connection");
+      return -1;
+    }
 
   fd = mkstemp (filename);
   FILE *file = fdopen (fd, "w");